pyproject.toml
[project]
name = "zoho-crm-mcp"
version = "0.1.1"
description = "Model Context Protocol server for Zoho CRM integration"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
zoho-mcp = "zoho_mcp.server:main"
zoho-mcp-auth = "zoho_mcp.auth_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["zoho_mcp"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]